requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
python-dotenv==1.0.1
pdf2image==1.17.0
Pillow==10.4.0
